Update NonQueryJinqStream.java
When the query compiler can not map a lambda to a native query and switches to the pure Java implementation, a distinct() in the NonQueryJinqStream may throw a StackOverflowError. distinct method is effectively calling itself before making any changes, so it seems like it should call super implementation, like many others.